+++ src/host_alerts/CountriesContactsAlert.cpp
@@ -2,7 +2,7 @@
 
 /* ***************************************************** */
 
-CountriesContactsAlert::CountriesContactsAlert(HostCheck *c, Host *f, risk_percentage cli_pctg, u_int64_t _countries_contacts, u_int64_t _countries_contacts_threshold) : HostAlert(c, f, cli_pctg) {
+CountriesContactsAlert::CountriesContactsAlert(HostCheck *c, Host *f, risk_percentage cli_pctg, u_int8_t _countries_contacts, u_int8_t _countries_contacts_threshold) : HostAlert(c, f, cli_pctg) {
     countries_contacts = _countries_contacts;
     countries_contacts_threshold = _countries_contacts_threshold;
 }
